Duration is suddenly double

+++ Data/Tracker/APIResults/RedditResult.cs
@@ -103,7 +103,7 @@ namespace MopsBot.Data.Tracker.APIResults.Reddit
         public int width { get; set; }
         public string scrubber_media_url { get; set; }
         public string dash_url { get; set; }
-        public int duration { get; set; }
+        public double duration { get; set; }
         public string hls_url { get; set; }
         public bool is_gif { get; set; }
         public string transcoding_status { get; set; }
@@ -141,7 +141,7 @@ namespace MopsBot.Data.Tracker.APIResults.Reddit
         public int width { get; set; }
         public string scrubber_media_url { get; set; }
         public string dash_url { get; set; }
-        public int duration { get; set; }
+        public double duration { get; set; }
         public string hls_url { get; set; }
         public bool is_gif { get; set; }
         public string transcoding_status { get; set; }
